Update the Deck and Card attribute
