+++ love_letter/models/cards.py
@@ -1,4 +1,6 @@
 import abc
+import random
+from typing import List
 
 
 class Card(metaclass=abc.ABCMeta):
@@ -6,15 +8,14 @@ class Card(metaclass=abc.ABCMeta):
     There are some properties in a card
 
     +---------+----------+
-    | Level   | Name     |
+    | Value   | Name     |
     +---------+----------+
-    | Picture            | * Quantity
+    | Picture | Quantity |
     +--------------------+
     | Effect Description |
     +--------------------+
 
-
-    * level: how intimate is this card with the princess
+    * value: how intimate is this card with the princess
     * name: the name of this card
     * (not yet planned) picture:
     * (not yet planned) description: the effect of this card
@@ -39,8 +40,8 @@ class Card(metaclass=abc.ABCMeta):
 
 class GuardCard(Card):
     name = '衛兵'
-    level = 1
-    number = 5
+    value = 1
+    quantity = 5
 
     def execute_with_card(self, player: "Player", guessing_card: "Card"):
         for card in player.cards:
@@ -50,26 +51,26 @@ class GuardCard(Card):
 
 class PriestCard(Card):
     name = '神父'
-    level = 2
-    number = 2
+    value = 2
+    quantity = 2
 
 
 class BaronCard(Card):
     name = '男爵'
-    level = 3
-    number = 2
+    value = 3
+    quantity = 2
 
 
 class HandmaidCard(Card):
     name = '侍女'
-    level = 4
-    number = 2
+    value = 4
+    quantity = 2
 
 
 class PrinceCard(Card):
     name = '王子'
-    level = 5
-    number = 2
+    value = 5
+    quantity = 2
 
     def can_not_play(self, player: "Player"):
         return COUNTESS_CARD in player.cards
@@ -83,8 +84,8 @@ class PrinceCard(Card):
 
 class KingCard(Card):
     name = '國王'
-    level = 6
-    number = 1
+    value = 6
+    quantity = 1
 
     def can_not_play(self, player: "Player"):
         return COUNTESS_CARD in player.cards
@@ -92,8 +93,8 @@ class KingCard(Card):
 
 class CountessCard(Card):
     name = '伯爵夫人'
-    level = 7
-    number = 1
+    value = 7
+    quantity = 1
 
     def execute_with_card(self, player: "Player", card: "Card"):
         pass
@@ -101,8 +102,8 @@ class CountessCard(Card):
 
 class PrincessCard(Card):
     name = '公主'
-    level = 8
-    number = 1
+    value = 8
+    quantity = 1
 
     def execute_with_card(self, player: "Player", card: "Card"):
         player.out()
@@ -113,6 +114,45 @@ def find_card_by_name(name):
         if card.name == name:
             return card
     raise ValueError(f'Cannot find the card with name: {name}')
+
+
+class Deck:
+    def __init__(self):
+        self.cards: List["Card"] = []
+        self.remove_by_rule_cards: List["Card"] = []
+
+    def shuffle(self, player_num: int):
+        """
+        Reset deck cards and remove rule cards.
+        :param player_num: 2 ~ 4
+        :return:
+        """
+        self.cards = []
+        self.remove_by_rule_cards = []
+
+        if player_num == 2:
+            remove_cards_num = 3
+        elif player_num in [3, 4]:
+            remove_cards_num = 1
+        else:
+            raise ValueError("player number is not support")
+
+        for card in ALL_CARD_TYPES:
+            self.cards.extend([card for _ in range(card.quantity)])
+
+        random.shuffle(self.cards)
+        random.shuffle(self.cards)
+
+        for num in range(remove_cards_num):
+            self.remove_by_rule_cards.append(self.cards.pop(0))
+
+    def draw(self, player: "Player"):
+        """
+        Player draw the top card.
+        :param player:
+        :return:
+        """
+        player.cards.append(self.cards.pop(0))
 
 
 ALL_CARD_TYPES = [
